--- src/main/java/com/brachy84/mechtech/MTMaterials.java
@@ -1,5 +1,0 @@
-package com.brachy84.mechtech;
-
-public class MTMaterials {
-
-}
