+++ src/main/java/com/brachy84/mechtech/comon/MTMaterials.java
@@ -1,0 +1,5 @@
+package com.brachy84.mechtech.comon;
+
+public class MTMaterials {
+
+}
