+++ src/main/java/com/brachy84/mechtech/comon/jei/TorusBlockRecipeWrapper.java
@@ -1,0 +1,30 @@
+package com.brachy84.mechtech.comon.jei;
+
+//import gregicadditions.GAValues;
+/*
+public class TorusBlockRecipeWrapper implements IRecipeWrapper {
+
+    private final TorusBlock torusBlock;
+
+    public TorusBlockRecipeWrapper(TorusBlock torusBlock) {
+        this.torusBlock = torusBlock;
+    }
+
+    @Override
+    public void getIngredients(IIngredients ingredients) {
+        ingredients.setInput(VanillaTypes.ITEM, torusBlock.getItemBlock());
+    }
+
+    @Override
+    public void drawInfo(Minecraft minecraft, int recipeWidth, int recipeHeight, int mouseX, int mouseY) {
+        drawLine(minecraft, "mechtech.recipe.torus_block.loss", 0, 18, 0x111111, torusBlock.getConductivity());
+        drawLine(minecraft, "mechtech.recipe.torus_block.range", 0, 27, 0x111111, torusBlock.getRangeModifier());
+        //drawLine(minecraft, "mechtech.recipe.torus_block.voltage", 0, 36, 0x111111, GAValues.VN[torusBlock.getMaxVoltageTier()], GAValues.V[torusBlock.getMaxVoltageTier()]);
+    }
+
+    public void drawLine(Minecraft minecraft, String langKey, int x, int y, int color, Object... obj) {
+        minecraft.fontRenderer.drawString(I18n.format(langKey, obj), x, y, color);
+    }
+
+}
+*/
