+++ src/main/java/com/brachy84/mechtech/comon/recipes/Recipes.java
@@ -1,0 +1,10 @@
+package com.brachy84.mechtech.comon.recipes;
+
+import com.brachy84.mechtech.comon.recipes.recipes.MetaTileEntityRecipes;
+
+public class Recipes {
+
+    public static void init() {
+        MetaTileEntityRecipes.init();
+    }
+}
