+++ src/main/java/com/brachy84/mechtech/comon/recipes/recipes/MetaTileEntityRecipes.java
@@ -1,0 +1,19 @@
+package com.brachy84.mechtech.comon.recipes.recipes;
+
+import com.brachy84.mechtech.comon.MTConfig;
+
+public class MetaTileEntityRecipes {
+
+    public static void init() {
+        if(MTConfig.multis.teslaTower.enabled) {
+            /*ASSEMBLER_RECIPES.recipeBuilder().duration(1200).EUt(2048)
+                    .input(gtMetalCasing, TungstenTitaniumCarbide, 4)
+                    .inputs(WIRE_COIL.getItemVariant(BlockWireCoil.CoilType.HSS_G, 16))
+                    .input("circuitSuperconductor", 4)
+                    .input(wireGtDouble, LuVSuperconductor, 32)
+                    .fluidInputs(SolderingAlloy.getFluid(1296))
+                    .outputs(TESLA_TOWER.getStackForm())
+                    .buildAndRegister();*/
+        }
+    }
+}
