--- src/main/java/com/brachy84/mechtech/recipes/Recipes.java
@@ -1,10 +1,0 @@
-package com.brachy84.mechtech.recipes;
-
-import com.brachy84.mechtech.recipes.recipes.MetaTileEntityRecipes;
-
-public class Recipes {
-
-    public static void init() {
-        MetaTileEntityRecipes.init();
-    }
-}
